Added /login endpoint to verify users
Created a login endpoint that will use basic authentication to verify
user to access to our application. From the request authorization header we grab the user name and password to verify against MySql databases Auth table. Taking email as username and a password.
With valid credentials, the user is returned a JWT to access the application.