Validate JWT Token
Added Validate endpoint that sends a POST request from the client
containing the users token in the Authorization header after being
validated as a user. To ensure the token hasnt been tampered with and/or
is valid we use JWT module to decode using secret_key and hashing
algorithm and returning status code 200.